[level 1] Title: x만큼 간격이 있는 n개의 숫자, Time: 4.30 ms, Memory: 82.9 MB -BaekjoonHub

+++ 프로그래머스/1/12954. x만큼 간격이 있는 n개의 숫자/x만큼 간격이 있는 n개의 숫자.java
@@ -4,7 +4,7 @@ import java.util.stream.*;
 class Solution {
     public long[] solution(int x, int n) {
         return LongStream.rangeClosed(1, n)
-                         .map(i -> (long)x * i)
+                         .map(i -> x * i)
                          .toArray();
     }
 }
